+++ outbound/leastping.go
@@ -2,10 +2,10 @@ package outbound
 
 import (
 	"github.com/sagernet/sing-box/adapter"
-	"github.com/sagernet/sing-box/balancer"
 	C "github.com/sagernet/sing-box/constant"
 	"github.com/sagernet/sing-box/log"
 	"github.com/sagernet/sing-box/option"
+	"github.com/sagernet/sing-box/outbound/balancer"
 )
 
 var (
